Make the bounds slightly larger to take in to account cluster size
Supercluster uses kdtrees which indexes points only, not rectangles.
Therefore the only option to capture all visible clusters is to pad the
search boundary with the half width/height of the cluster.

I removed cluster size calculation since it would not be possible to
determine the half width/height height of the displayed clusters before
actually searching them in order to pad the search area.